Create common class with its tests
