+++ test/unit/webdriver/device/common_test.py
@@ -32,3 +32,13 @@ class TestWebDriverNotifications(object):
             appium_command('/session/1234567890/appium/device/open_notifications')
         )
         assert isinstance(driver.open_notifications(), WebDriver)
+
+    @httpretty.activate
+    def test_current_package(self):
+        driver = android_w3c_driver()
+        httpretty.register_uri(
+            httpretty.GET,
+            appium_command('/session/1234567890/appium/device/current_package'),
+            body='{"value": ".ExamplePackage"}'
+        )
+        assert driver.current_package == '.ExamplePackage'
